Refactor Paddlepaddle and Onnx tests and Remove all Pytorch and JAX Tests (#3211)
### Ticket
Fixes #3183 

### Problem description
Load all models and inputs from tt-forge-models (via its loaders),
instead of loading PyTorch models in-test for mentioned onnx and
paddlepaddle models.

### What's changed

- Refactor onnx tests for below models with created loaders:

     - AlexNet

     -  DenseNet

     -  Efficientnet
  
     -   Googlenet
  
     -  MobileNetV1
  
     - Resnet

- Refactor paddlepaddle tests for below models with created loaders:

     -  AlexNet

     - DenseNet

     - Googlenet

     -  MobileNetV2

     - Resnet

- Tracked latest main of tt_forge_models and fixed the regressions
caused by it.

- Removed all pytorch tests and Jax tests.

### Logs

[alexnet_onnx_forge.log](https://github.com/user-attachments/files/26043192/alexnet_onnx_forge.log)

[googlenet_onnx_forge.log](https://github.com/user-attachments/files/26043193/googlenet_onnx_forge.log)

[densenet_onnx_forge.log](https://github.com/user-attachments/files/26043196/densenet_onnx_forge.log)

[res_onnx_forge.log](https://github.com/user-attachments/files/26043198/res_onnx_forge.log)

[mobilenetv1_onnx_forge.log](https://github.com/user-attachments/files/26133290/mobilenetv1_onnx_forge.log)

[alexnet_paddle_forge.log](https://github.com/user-attachments/files/26133307/alexnet_paddle_forge.log)

[densenet_paddle_forge.log](https://github.com/user-attachments/files/26133311/densenet_paddle_forge.log)

[googlenet_paddle_forge.log](https://github.com/user-attachments/files/26133312/googlenet_paddle_forge.log)

[mobilenetv2_paddle_forge.log](https://github.com/user-attachments/files/26133326/mobilenetv2_paddle_forge.log)

[resnet_padle_forge.log](https://github.com/user-attachments/files/26133340/rensnet_padle_forge.log)

---------

+++ forge/test/models/onnx/vision/yolo/model_utils/yolo_utils.py
@@ -5,9 +5,8 @@
 import os
 import torch
 from torchvision import transforms
-from typing import Optional
 from ultralytics.nn.tasks import DetectionModel
-
+from typing import Optional
 from test.utils import download_model
 from third_party.tt_forge_models.tools.utils import get_file
 
@@ -23,7 +22,9 @@ class YoloWrapper(torch.nn.Module):
 
     def forward(self, image: torch.Tensor):
         y, x = self.model(image)
-        return (y, *x)
+        # Post processing inside model casts output to float32, even though raw output is aligned with image.dtype
+        # Therefore we need to cast it back to image.dtype
+        return (y.to(image.dtype), *x)
 
 
 def _create_image_input() -> torch.Tensor:
